refactor: componentes do dashboard.
