+++ src/views/Dashboard/Dashboard.spec.js
@@ -1,13 +1,35 @@
 import { describe, expect, it, vi } from 'vitest';
 import { mount } from '@vue/test-utils'
 import Dashboard from './Dashbord.vue'
+
 import DashboardAdmin from '../../components/DashboardAdmin.vue'
 import DashboardReceptionist from '../../components/DashboardReceptionist.vue'
 import DashboardInstructor from '../../components/DashboardInstructor.vue'
 import DashboardNutritionist from '../../components/DashboardNutritionist.vue'
 import DashboardStudent from '../../components/DashboardStudent.vue'
 
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+const vuetify = createVuetify({
+  components,
+  directives
+})
+
+global.ResizeObserver = require('resize-observer-polyfill')
+
 describe('Dashboard Component', () => {
+
+  it('Espera-se que a tela Dashboard seja renderizada', () => {
+    const component = mount(Dashboard, {
+      global: {
+        plugins: [vuetify]
+      }
+    })
+    expect(component).toBeTruthy()
+  })
+
   it('Deve renderizar o componente DashboardAdmin se o perfil for ADMIN', async () => {
     const localStorageMock = {
       getItem: vi.spy().andReturn('ADMIN')
